:recycle: Rename to data package
