+++ data/check_all_case_when_exported_package.go
@@ -1,4 +1,4 @@
-package testdata
+package data
 
 import "./x"
