+++ data/check_not_enum_case_function.go
@@ -1,4 +1,4 @@
-package testdata
+package data
 
 type language1 int
