+++ data/check_wrote_all_case_function.go
@@ -1,4 +1,4 @@
-package testdata
+package data
 
 type language2 int
